requirements.txt
pyqt5
ffmpeg-python
opencv-python
